test: split use-find into tests by paginateOn type
